remove tauri dependency from the devtools crate
