+++ devtools/Cargo.toml
@@ -18,16 +18,15 @@ tracing-subscriber.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["macros", "time"] }
 tokio-stream = { workspace = true, features = ["fs"] }
-tauri.workspace = true
 prost-types.workspace = true
 tonic.workspace = true
 tonic-web = "0.10"
 tonic-health = "0.10"
 tower-http = "0.4"
+tower-layer = "0.3"
 futures = "0.3.28"
 bytes = "1.5.0"
 ringbuf = "0.4.0-rc.2"
-colored = "2.0"
 async-stream = "0.3.5"
 http = "0.2"
