+++ tauri-plugin/Cargo.toml
@@ -15,4 +15,9 @@ tauri.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
 tokio.workspace = true
+tonic.workspace = true
 colored = "2.0"
+serde_json = "1"
+futures = "0.3.28"
+bytes = "1.5.0"
+async-stream = "0.3.5"
