+++ wire/Cargo.toml
@@ -15,8 +15,6 @@ tonic.workspace = true
 prost.workspace = true
 prost-types.workspace = true
 tracing-core.workspace = true
-tauri.workspace = true
-serde_json = "1.0"
 bitflags = "2.4"
 
 [dev-dependencies]
