+++ wire/src/tauri.rs
@@ -5,12 +5,3 @@ mod generated {
 }
 
 pub use generated::*;
-
-// TODO improve this
-impl<'a> From<&'a tauri::Config> for Config {
-    fn from(value: &'a tauri::Config) -> Self {
-        Self {
-            raw: serde_json::to_string(&value).unwrap(),
-        }
-    }
-}
